fix: update noticeService
