+++ backend/src/main/java/com/varchar6/petcast/common/response/ResponseMessage.java
@@ -2,9 +2,6 @@ package com.varchar6.petcast.common.response;
 
 
 import lombok.*;
-import org.springframework.http.HttpStatus;
-
-import java.util.Map;
 
 @NoArgsConstructor
 @AllArgsConstructor
